Trying to count scores

+++ backend/kendo-tournament-core/src/main/java/com/softwaremagico/kt/core/controller/models/FightStatisticsDTO.java
@@ -25,17 +25,17 @@ package com.softwaremagico.kt.core.controller.models;
  */
 
 public class FightStatisticsDTO extends ElementDTO {
-    private Integer fightsNumber;
+    private Long fightsNumber;
     private Integer fightsByTeam;
-    private Integer duelsNumber;
+    private Long duelsNumber;
     //In seconds.
     private Long time;
 
-    public Integer getFightsNumber() {
+    public Long getFightsNumber() {
         return fightsNumber;
     }
 
-    public void setFightsNumber(Integer fightsNumber) {
+    public void setFightsNumber(Long fightsNumber) {
         if (fightsNumber != null && fightsNumber >= 0) {
             this.fightsNumber = fightsNumber;
         } else {
@@ -43,11 +43,11 @@ public class FightStatisticsDTO extends ElementDTO {
         }
     }
 
-    public Integer getDuelsNumber() {
+    public Long getDuelsNumber() {
         return duelsNumber;
     }
 
-    public void setDuelsNumber(Integer duelsNumber) {
+    public void setDuelsNumber(Long duelsNumber) {
         if (duelsNumber != null && duelsNumber >= 0) {
             this.duelsNumber = duelsNumber;
         } else {
